Delegate addV3ClientNamedImport downstream
